[Mac] Fix daemonization under MacOS.
In macOS, an executable from the same app bundle
can't have two instances, but for spirit we need two
instances of the our app bundle, one is the daemon and
another is the control app. So to solve this we copy our
own app bundle to a hidden directory and then execute it
from the control app.