[Refactor/Bugfix] Condense Active Window Tracker.
Active Window Tracker under X11 is a little bit buggy
since KWindowSystem signals are just too fast which is my
assumption that sometimes Spirit is killed when the window
it's tracking is killed. We now have some checks to avoid this and
introduced some delay, maybe add more if needed.

Also we made Active Window Tracker a little bit lightweight by using
the same mechanism for all platforms which seems to make sense.

Signed-off-by: Divya Antony J R <antonyjr@protonmail.com>